Moved LRULinkedHashMap to utils
