+++ java_src/foam/nanos/auth/CachedUserAndGroupAuthService.java
@@ -1,6 +1,8 @@
 package foam.nanos.auth;
 
 import foam.core.X;
+import foam.nanos.util.LRULinkedHashMap;
+
 import java.util.Map;
 
 /**
